Crate slapi_r_plugin

Modules

constants
Constants Module

entry
Entry module.

error
Error Module

log
Log Modules

pblock
Pblock Module

plugin
Plugin Module

Macros

slapi_r_plugin_init!
This macro is imported by plugins to wrap their rust init functions. This allows complete transparency to C types to the rust plugin.