Module slapi_r_plugin::constants

Constants Module

This module defines the Directory Server library constants that are defined in slapi-public.h and slapi-plugin.h.

Enums

LogLevel
Error logging levels that may be used. This is controlled by -d on the ns-slapd commandline.

PluginVersion
Plugin API versions that can be used for a plugin. You should in most cases use version _03

Constants

LDAP_SUCCESS
A successful operation. Most Directory Server fuctions expect this.

SLAPI_PLUGIN_BE_TXN_POST_ADD_FN
PBlock constant for registering a post add operation.

SLAPI_PLUGIN_BE_TXN_POST_DELETE_FN
PBlock constant for registering a post delete operation.

SLAPI_PLUGIN_BE_TXN_POST_MODIFY_FN
PBlock constant for registering a post modify operation.

SLAPI_PLUGIN_BE_TXN_POST_MODRDN_FN
PBlock constant for registering a post modrdn operation.

SLAPI_PLUGIN_BE_TXN_PRE_ADD_FN
PBlock constant for registering a pre BE_TXN add operation.

SLAPI_PLUGIN_BE_TXN_PRE_DELETE_FN
PBlock constant for registering a pre BE_TXN delete operation.

SLAPI_PLUGIN_BE_TXN_PRE_DELETE_TOMBSTONE_FN
PBlock constant for registering a pre BE_TXN delete tombstone operatation.

SLAPI_PLUGIN_BE_TXN_PRE_MODIFY_FN
PBlock constant for registering a pre BE_TXN modify operation.

SLAPI_PLUGIN_BE_TXN_PRE_MODRDN_FN
PBlock constant for registering a pre BE_TXN modrdn operation.

SLAPI_PLUGIN_CLOSE_FN
PBlock constant for registering the close function for a plugin.

SLAPI_PLUGIN_DESCRIPTION
PBlock constant for accessing plugin description information.

SLAPI_PLUGIN_OPRETURN
PBlock constant for accessing plugin operation return codes.

SLAPI_PLUGIN_POST_SEARCH_FN
PBlock constant for registering a post search operation.

SLAPI_PLUGIN_PRIVATE
PBlock constant for accessing private plugin data.

SLAPI_PLUGIN_START_FN
PBlock constant for registering the start function for a plugin.

SLAPI_PLUGIN_TYPE
PBlock constant for accessing plugin type.

SLAPI_PLUGIN_VERSION
PBlock constant for accessing plugin API version.

SLAPI_SEARCH_RESULT_ENTRY
PBlock constant for retrieving the next entry of an entry result list.