Struct slapi_r_plugin::entry::Slapi_R_Entry

pub struct Slapi_R_Entry {
    // some fields omitted
}

Slapi_R_Entry is a container for a slapi_entry C type. This represents an entry that has been retrieved from the revelant backend as part of an operation. By making this opaque, we create rewrite and improvement possibilities

Methods

impl Slapi_R_Entry

fn new(slapi_entry: *const c_void) -> Slapi_R_Entry

Build a new Slapi_R_Entry from a pointer to the slapi_entry The slapi_entry is created in a number of paths through DS, so you should not need to do this yourself.