Module slapi_r_plugin::error

Error Module

This module contains definitions of all the possible Error Results that Directory Server or a plugin may return. It is used to facilitate conversion of errors between C and Rust in native types, as well as allowing detailed and specific error types to be created in Rust as Super types.

Enums

LoggingError
This type represents an error while send a log to Directory server This error may be from Directory Server and is returned to the Rust code.

PBlockError
This type represents an error while working with the Slapi_R_PBlock. This error may be from Directory Server and is returned to the Rust code.

PluginOperationError
This type is used by a plugin operation Ie post_search, to represent a failure during the processing. It is returned to Directory Server.

PluginRegistrationError
This type is used when a plugin is being registered with directory server It represents an error in the plugin that should be returned.