Module slapi_r_plugin::pblock

Pblock Module

This module wraps the Directory Server Parameter Block data structure. The Slapi_PBlock is an important datastructure, used for nearly all operations in the Directory Server. Requests and Responses are placed into it. This module abstracts the Slapi_PBlock into Rust types, and wraps the get and set functions into various types. This allows each type to be changed in the future without affecting the stability of plugins.

Structs

Slapi_R_PBlock
Slapi_R_PBlock is a container that contains the slapi_pblock C type. Functions of Slapi_R_PBlock wrap and interact with the native C PBlock. By making this opaque, it gives us the ability to rework or improve the structure without breaking the API of Slapi_R_PBlock.