Struct slapi_r_plugin::pblock::Slapi_R_PBlock

pub struct Slapi_R_PBlock {
    // some fields omitted
}

Slapi_R_PBlock is a container that contains the slapi_pblock C type. Functions of Slapi_R_PBlock wrap and interact with the native C PBlock. By making this opaque, it gives us the ability to rework or improve the structure without breaking the API of Slapi_R_PBlock.

Methods

impl Slapi_R_PBlock

fn new() -> Slapi_R_PBlock

Creates a new Slapi_R_PBlock, with a new C PBlock structure.

Panics

Directory Server's pblock functions DO NOT check for null pointer deref It is possible while using a new Slapi_R_PBlock to trigger a SIGSEGV as pblock internal structures such as pb_plugin are not allocated.

fn build(slapi_pblock: *const c_void) -> Slapi_R_PBlock

Builds a new Slapi_R_PBlock from a Slapi_PBlock C pointer. It is very likely, that this is the function you want to use if you are woring on slapi_r_plugin internals.

fn init(self)

This will wipe and reset the internal contents of the Slapi_PBlock C structure.

fn destroy(self)

This will free and deallocate all parts o fthe Slapi_PBlock C structure. It is the last thing you should use.

Panics

After calling this function, if you access any other function of this struct, you will likely cause a SIGSEGV

fn get_plugin_type(&self) -> Option<isize>

This will retrieve the value of SLAPI_PLUGIN_TYPE, such as BE_TXN, POST_OP etc.

fn set_plugin_type(&self, plugin_type: isize)

This will set the plugin type value as SLAPI_PLUGIN_TYPE, such as BE_TXN, POST_OP, etc.

fn get_plugin_opreturn(&self) -> Option<isize>

This will get the operation return code as SLAPI_PLUGIN_OPRETURN

fn set_plugin_opreturn(&self, opreturn: isize)

This will set the operation return code as SLAPI_PLUGIN_OPRETURN

fn get_plugin_version(&self) -> Option<isize>

This will get the plugin api version from SLAPI_PLUGIN_VERSION See also constants::PluginVersion

fn set_plugin_version(&self, version: isize)

This will set the plugin api version from SLAPI_PLUGIN_VERSION See also constants::PluginVersion

fn set_plugin_close_fn(&self, func: extern "C" fn(*const c_void) -> isize)

This will set the close plugin callback handler as SLAPI_PLUGIN_CLOSE_FN. You should not call this directly as the Slapi_R_Plugin_Manager will handle this for you.

fn set_plugin_start_fn(&self, func: extern "C" fn(*const c_void) -> isize)

This will set the start plugin callback handler as SLAPI_PLUGIN_START_FN. You should not call this directly as the Slapi_R_Plugin_Manager will handle this for you.

fn set_plugin_post_search_fn(&self, func: extern "C" fn(*const c_void) -> isize)

This will set the post search operation plugin callback handler as SLAPI_PLUGIN_POST_SEARCH_FN. You should not call this directly as the Slapi_R_Plugin_Manager will handle this for you.

fn get_plugin_private<T>(&self) -> Option<&T>

This will get a pointer to a structure stored in the plugin private data stash, from Slapi_PBlock.pb_plugin->plg_private. SLAPI_PLUGIN_PRIVATE You should never call this directly, as certain parts of the Slapi_R_Plugin_Manager rely on this data being un-tampered.

fn set_plugin_private<T>(&self, value: T)

This will memory copy and set a pointer to the memory of a generic Rust type into a 'box' stored in the Slapi_PBlock for this plugin instance. Certain parts of Slapi_R_Plugin_Manager rely on this, so you should never call this directly.

fn destroy_plugin_private(&self) -> Result<(), PBlockError>

This will destroy (free) the contents of the pointer stored within the Slapi_PBlock instance. Certain parts of Slapi_R_Plugin_Manager rely on this, so you should never call this directly.

fn get_search_result_entry(&self) -> Option<Slapi_R_Entry>

This will retrieve the next Slapi_R_Entry from the result set in the pblock.