Module slapi_r_plugin::plugin

Plugin Module

This module is responsible for wrapping and proxying the interactions plugins and Directory Server. The motivation is to simplify the design and creation of plugins in Rust, so that they do not require exposure to the details of the C types that Directory Server provides.

Structs

Slapi_R_Plugin_FN
Type that represents the possible call backs from a plugin. This is passed into slapi_pblock-...->plg_private, so that when our Regisitered proxy functions are called, we are able to look up and access the correct rust function call backs.

Slapi_R_Plugin_Manager
Type that represents a plugin that we will register. After registration this instance is destroyed.