Struct slapi_r_plugin::plugin::Slapi_R_Plugin_FN

pub struct Slapi_R_Plugin_FN {
    pub start: Option< extern "C" fn(&Slapi_R_PBlock) -> Result<(), PluginOperationError>>,
    pub close: Option< extern "C" fn(&Slapi_R_PBlock) -> Result<(), PluginOperationError>>,
    pub post_search: Option< extern "C" fn(&Slapi_R_PBlock) -> Result<(), PluginOperationError>>,
}

Type that represents the possible call backs from a plugin. This is passed into slapi_pblock-...->plg_private, so that when our Regisitered proxy functions are called, we are able to look up and access the correct rust function call backs.

Fields

start

On option type for a function callback that handles plugin start up.

close

On option type for a function callback that handles plugin close down.

post_search

An option type for a function callback that handles post search.