Struct slapi_r_plugin::plugin::Slapi_R_Plugin_Manager

pub struct Slapi_R_Plugin_Manager<'a> {
    pub name: &'a str,
    pub version: PluginVersion,
    pub functions: Slapi_R_Plugin_FN,
}

Type that represents a plugin that we will register. After registration this instance is destroyed.

Fields

name

The name of the plugin.

version

The slapi plugin version api that is provided by the plugin. Defaults to 3

functions

The set of optional callbacks that the plugin registers.

Methods

impl<'a> Slapi_R_Plugin_Manager<'a>

fn new() -> Slapi_R_Plugin_Manager<'a>

Builds a new Slapi_R_Plugin_Manager. The Rust plugin can then set values on this struct, and finally will call .register() to complete the plugins initialisation.

fn register(self, pb: Slapi_R_PBlock) -> Result<(), PluginRegistrationError>

Completes the registration to Directory Server of the plugin. This is the last function you call when building a plugin in a plugin init.